examples: add examples to trilinos workflow
